AlpacaTutorial/Alpaca28/requirements.txt
alpaca-py==0.29.2


